+++ mobile/android/android-components/components/browser/search/src/main/java/mozilla/components/browser/search/provider/AssetsSearchEngineProvider.kt
@@ -15,6 +15,7 @@ import mozilla.components.browser.search.SearchEngineParser
 import mozilla.components.browser.search.provider.filter.SearchEngineFilter
 import mozilla.components.browser.search.provider.localization.SearchLocalizationProvider
 import mozilla.components.support.ktx.android.content.res.readJSONObject
+import mozilla.components.support.ktx.android.org.json.toList
 import org.json.JSONArray
 import org.json.JSONObject
 
@@ -30,6 +31,7 @@ import org.json.JSONObject
  * Optionally <code>additionalIdentifiers</code> to be loaded can be specified. A search engine
  * identifier corresponds to the search plugin XML file name (e.g. duckduckgo -> duckduckgo.xml).
  */
+@Suppress("TooManyFunctions")
 class AssetsSearchEngineProvider(
     private val localizationProvider: SearchLocalizationProvider,
     private val filters: List<SearchEngineFilter> = emptyList(),
@@ -40,13 +42,37 @@ class AssetsSearchEngineProvider(
     /**
      * Load search engines from this provider.
      */
-    override suspend fun loadSearchEngines(context: Context): List<SearchEngine> {
-        val searchEngineIdentifiers = mutableListOf<String>().apply {
-            addAll(loadAndFilterConfiguration(context))
-            addAll(additionalIdentifiers)
-        }
+    override suspend fun loadSearchEngines(context: Context): SearchEngineList {
+        val localizedConfiguration = loadAndFilterConfiguration(context)
+        val searchEngineIdentifiers = localizedConfiguration.visibleDefaultEngines + additionalIdentifiers
 
-        return loadSearchEnginesFromList(context, searchEngineIdentifiers.distinct())
+        val searchEngines = loadSearchEnginesFromList(
+            context,
+            searchEngineIdentifiers.distinct()
+        )
+
+        // Reorder the list of search engines according to the configuration.
+        // Note: we're using the name of the search engine, not the id, so we can only do this
+        // after we've loaded the search engine from the XML
+        val searchOrder = localizedConfiguration.searchOrder
+        val orderedList = searchOrder
+            .mapNotNull { name ->
+                searchEngines.find { it.name == name }
+            }
+
+        val unorderedRest = searchEngines
+            .filter {
+                !searchOrder.contains(it.name)
+            }
+
+        val defaultEngine = localizedConfiguration.searchDefault?.let { name ->
+                searchEngines.find { it.name == name }
+            }
+
+        return SearchEngineList(
+            orderedList + unorderedRest,
+            defaultEngine
+        )
     }
 
     private suspend fun loadSearchEnginesFromList(
@@ -92,19 +118,26 @@ class AssetsSearchEngineProvider(
         identifier: String
     ): SearchEngine = parser.load(assets, identifier, "searchplugins/$identifier.xml")
 
-    private fun loadAndFilterConfiguration(context: Context): List<String> {
+    private fun loadAndFilterConfiguration(context: Context): SearchEngineListConfiguration {
         val config = context.assets.readJSONObject("search/list.json")
 
-        val configBlock = pickConfigurationBlock(config)
-        val jsonSearchEngineIdentifiers = getSearchEngineIdentifiersFromBlock(configBlock)
+        val configBlocks = pickConfigurationBlocks(config)
+        val jsonSearchEngineIdentifiers = getSearchEngineIdentifiersFromBlock(configBlocks)
 
-        return applyOverridesIfNeeded(config, jsonSearchEngineIdentifiers)
+        val searchOrder = getSearchOrderFromBlock(configBlocks)
+        val searchDefault = getSearchDefaultFromBlock(configBlocks)
+
+        return SearchEngineListConfiguration(
+            applyOverridesIfNeeded(config, jsonSearchEngineIdentifiers),
+            searchOrder.toList(),
+            searchDefault
+        )
     }
 
-    private fun pickConfigurationBlock(config: JSONObject): JSONObject {
+    private fun pickConfigurationBlocks(config: JSONObject): Array<JSONObject> {
         val localesConfig = config.getJSONObject("locales")
 
-        return when {
+        val localizedConfig = when {
             // First try (Locale): locales/xx_XX/
             localesConfig.has(localizationProvider.languageTag) ->
                 localesConfig.getJSONObject(localizationProvider.languageTag)
@@ -113,18 +146,60 @@ class AssetsSearchEngineProvider(
             localesConfig.has(localizationProvider.language) ->
                 localesConfig.getJSONObject(localizationProvider.language)
 
-            // Third try: Use default
-            else -> config
+            // Give up, and fallback to defaults
+            else -> null
         }
+
+        return localizedConfig?.let {
+            arrayOf(it, config)
+        } ?: arrayOf(config)
     }
 
-    private fun getSearchEngineIdentifiersFromBlock(configBlock: JSONObject): JSONArray {
+    private fun getSearchEngineIdentifiersFromBlock(configBlocks: Array<JSONObject>): JSONArray {
         // Now test if there's an override for the region (if it's set)
-        return if (localizationProvider.region != null && configBlock.has(localizationProvider.region)) {
-            configBlock.getJSONObject(localizationProvider.region)
-        } else {
-            configBlock.getJSONObject("default")
-        }.getJSONArray("visibleDefaultEngines")
+        return getArrayFromBlock("visibleDefaultEngines", configBlocks)
+            ?: throw IllegalStateException("No visibleDefaultEngines for " +
+                    "${localizationProvider.languageTag} in " +
+                    (localizationProvider.region ?: "default"))
+    }
+
+    private fun getSearchDefaultFromBlock(configBlocks: Array<JSONObject>): String? {
+        return getStringFromBlock("searchDefault", configBlocks)
+    }
+
+    private fun getSearchOrderFromBlock(configBlocks: Array<JSONObject>): JSONArray? {
+        return getArrayFromBlock("searchOrder", configBlocks)
+    }
+
+    private fun getStringFromBlock(key: String, blocks: Array<JSONObject>): String? =
+            getValueFromBlock(blocks) {
+                it.optString(key, null)
+            }
+
+    private fun getArrayFromBlock(key: String, blocks: Array<JSONObject>): JSONArray? =
+            getValueFromBlock(blocks) {
+                it.optJSONArray(key)
+            }
+
+    /**
+     * This looks for a JSONObject in the config blocks it is passed that is able to be transformed
+     * into a value. It tries the permutations of locale and region specific from most specific to
+     * least specific.
+     *
+     * This has to be done on a value basis, not a configBlock basis, as the configuration for a
+     * given locale/region is not grouped into one object, but spread across the json file,
+     * according to these rules.
+     */
+    private fun <T : Any> getValueFromBlock(blocks: Array<JSONObject>, transform: (JSONObject) -> T?): T? {
+        val regions = arrayOf(localizationProvider.region, "default")
+            .mapNotNull { it }
+
+        return blocks
+            .flatMap { block ->
+                regions.mapNotNull { region -> block.optJSONObject(region) }
+            }
+            .mapNotNull(transform)
+            .firstOrNull()
     }
 
     private fun applyOverridesIfNeeded(config: JSONObject, jsonSearchEngineIdentifiers: JSONArray): List<String> {
@@ -147,3 +222,9 @@ class AssetsSearchEngineProvider(
         return searchEngineIdentifiers
     }
 }
+
+internal data class SearchEngineListConfiguration(
+    val visibleDefaultEngines: List<String>,
+    val searchOrder: List<String>,
+    val searchDefault: String?
+)