+++ mobile/android/android-components/components/browser/search/src/main/java/mozilla/components/browser/search/provider/SearchEngineProvider.kt
@@ -14,5 +14,22 @@ interface SearchEngineProvider {
     /**
      * Load search engines from this provider.
      */
-    suspend fun loadSearchEngines(context: Context): List<SearchEngine>
+    suspend fun loadSearchEngines(context: Context): SearchEngineList
 }
+
+/**
+ * Data class providing an ordered list of search engines and a default search engine from a
+ * specific source.
+ */
+data class SearchEngineList(
+
+    /**
+     * An ordered list of search engines.
+     */
+    val list: List<SearchEngine>,
+
+    /**
+     * The default search engine if the user has no preference.
+     */
+    val default: SearchEngine?
+)